Reworked User_data
I worked on the new user page so the submit button uploads the first and last name information to the database, I linked the authenticated user with a certain User_data array so the Id of an authenticated user can pull all the relevant information you would need. I Redid the event details page so that it displays all the necessary information for an event. I reworked the Home page so that it appears closer to our mockup. I also limited the event cards so the information doesn't overflow. I created a color theme file where we can now import colors and use them. I fixed the weird bug with the dates adding an hour.